Move control schema to schemas; drop schema helper
 - Extract _control_schema from config_flow into schemas.py and import it there.
 - Remove the schema_with_defaults helper from util.py.
 - Update config_flow to use _device_schema()/ _device_schema(defaults) for form rendering and validation.
 - Fix device item lookups for primary/secondary to use safe defaults. Also adjust imports.
