+++ custom_components/tesira_ttp/util.py
@@ -1,8 +1,8 @@
 # #################################################################################################################### #
 # Filename: \custom_components\tesira_ttp\util.py                                                                      #
 # Repository: tesira_ttp                                                                                               #
-# Created Date: Saturday, March 28th 2026, 10:45:20 PM                                                                 #
-# Last Modified: Friday, April 3rd 2026, 9:21:44 PM                                                                    #
+# Created Date: Thursday, March 19th 2026, 12:56:52 AM                                                                 #
+# Last Modified: Sunday, April 5th 2026, 9:19:42 PM                                                                    #
 # Original Author: Darnel Kumar                                                                                        #
 # Author Github: https://github.com/Darnel-K                                                                           #
 #                                                                                                                      #
@@ -26,50 +26,10 @@
 
 from __future__ import annotations
 
-from typing import Any, Dict
+from typing import Dict
 
-import voluptuous as vol
 import base64
 import json
-
-# =====================================================================
-# Schema Helpers
-# =====================================================================
-def schema_with_defaults(
-    base_schema: vol.Schema, defaults: Dict[str, Any]
-) -> vol.Schema:
-    """
-    Create a new voluptuous Schema with runtime default values injected.
-
-    This **does not** modify the passed-in base schema.
-
-    Args:
-        base_schema: The original schema to copy fields from.
-        defaults: A dictionary mapping field names -> default values.
-
-    Returns:
-        A new vol.Schema instance where matching fields are replaced with
-        new Required/Optional fields containing default values.
-    """
-    new_fields: Dict[Any, Any] = {}
-
-    # base_schema.schema contains {vol.Required/vol.Optional: validator}
-    for key, validator in base_schema.schema.items():
-        field_name = key.schema  # the raw field name string
-
-        if field_name in defaults:
-            default_value = defaults[field_name]
-
-            # Replace field with one that has a default applied
-            if isinstance(key, vol.Required):
-                new_fields[vol.Required(field_name, default=default_value)] = validator
-            else:
-                new_fields[vol.Optional(field_name, default=default_value)] = validator
-        else:
-            # Leave untouched
-            new_fields[key] = validator
-
-    return vol.Schema(new_fields)
 
 def gen_hub_key(deviceModel: str, deviceRevision: int, serialNumber: str, format: str = "b64") -> str:
     """
